# docs/openapi.yaml
openapi: 3.1.0
info:
  title: coupon-api
  description: クーポン基盤の実行時統合API。会員×クーポンの事前紐づけを持たず、会員属性を受け取って該当クーポンを判定・返却する。
  version: 0.1.0
servers:
- url: http://localhost:8081
  description: Generated server url
...
paths:
...
  /internal/events:
    post:
      tags:
      - internal-events
      summary: ドメインイベント受領（冪等取り込み）
      description: coupon-admin の fat event を受領し、percolator＋クーポンマスタ投影を更新する。eventId
        重複排除＋version の last-writer-wins。不正イベントは適用せず 422（DLQ 相当）。
      operationId: receive
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponEventEnvelopeRequest"
        required: true
      responses:
        "202":
          description: 受領（APPLIED / DUPLICATE / STALE）
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventIngestResponse"
        "422":
          description: 不正イベント（適用せず DLQ）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /internal/events/resync:
    post:
      tags:
      - internal-events
      summary: 全件resync（スナップショット再適用）
      description: coupon-admin から再発行された現在状態のスナップショットで read model を作り直す。トリガー・頻度（手動/定期/イベント起動）は未確定（変更ブリーフ
        §4）。
      operationId: resync
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResyncRequest"
        required: true
      responses:
        "200":
          description: 再構築完了
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventIngestResponse"
  /coupons:
    post:
      tags:
      - coupons
      summary: クーポンマスタ登録（upsert）
      description: eligibility 判定のためのシードパス。同一 couponId は上書き。
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCouponRequest"
        required: true
      responses:
        "201":
          description: 登録済み
        "400":
          description: 入力検証・ドメイン検証エラー
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /coupons/eligibility:
    post:
      tags:
      - coupons
      summary: クーポン一覧取得（属性に基づく該当判定＋全員配信の付与。差し引きなしの生集合）
      description: "percolate 該当の SEGMENT（有効期間内）∪ 有効期間内の ALL の生集合を返す。消費済み・選択状態・SUSPENDED\
        \ の差し引きは行わない（会員向け表示一覧の合成は POST /members/{memberId}/coupon-list、BFF 経路の差し引き\
        は BFF の責務）。"
      operationId: eligibility
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EligibilityRequest"
        required: true
      responses:
        "200":
          description: 該当クーポン（ID＋distributionType）の一覧
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EligibilityResponse"
        "400":
          description: 属性・判定基準時刻の検証エラー
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    CouponEventDataRequest:
      type: object
      description: イベントペイロード。CouponDeleted は couponId のみ、それ以外は投影に必要な全項目。
      properties:
        couponId:
          type: string
          minLength: 1
        schemaVersion:
          type: string
          description: 条件スキーマ版（coupon-rule-schema）
          example: 1.0
        condition:
          description: 配布条件（正準AST・coupon-rule-schema 準拠）。SEGMENT のとき必須相当。
        distributionType:
          type: string
          enum:
          - ALL
          - SEGMENT
          - INDIVIDUAL
        effectiveFrom:
          type: string
          format: date-time
          description: 有効期間の開始（含む・RFC 3339）
          example: 2026-07-01T00:00:00+09:00
        effectiveTo:
          type: string
          format: date-time
          description: 有効期間の終了（含む・RFC 3339）
          example: 2026-07-31T23:59:59+09:00
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
      required:
      - couponId
    CouponEventEnvelopeRequest:
      type: object
      description: coupon-admin ドメインイベント（fat event / CloudEvents 風）
      properties:
        eventId:
          type: string
          description: イベントID（冪等の重複排除キー）
          example: evt-01H...
          minLength: 1
        type:
          type: string
          description: イベント型
          example: CouponApproved
          minLength: 1
        version:
          type: integer
          format: int64
          description: クーポンの版（last-writer-wins）
          example: 7
        occurredAt:
          type: string
          format: date-time
          description: 発生時刻（RFC 3339）
          example: 2026-07-09T09:00:00Z
        traceId:
          type: string
          description: W3C Trace Context の相関ID
          example: 0af7651916cd43dd8448eb211c80319c
        data:
          $ref: "#/components/schemas/CouponEventDataRequest"
      required:
      - data
      - eventId
      - occurredAt
      - type
      - version
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
    EventIngestResponse:
      type: object
      description: 取り込み結果（APPLIED / DUPLICATE / STALE / RESYNCED）
      properties:
        outcome:
          type: string
      required:
      - outcome
    ResyncRequest:
      type: object
      description: 全件resync リクエスト（現在状態を表すイベントのスナップショット）
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/CouponEventEnvelopeRequest"
      required:
      - events
    RegisterCouponRequest:
      type: object
      description: クーポンマスタ登録（upsert）リクエスト
      properties:
        couponId:
          type: string
          description: クーポンID
          example: CP-2026-0001
          minLength: 1
        distributionType:
          type: string
          description: 配布種別。ALL=全員配信 / SEGMENT=属性該当 / INDIVIDUAL=個別付与（属性では決まらない）
          enum:
          - ALL
          - SEGMENT
          - INDIVIDUAL
        validFrom:
          type: string
          format: date-time
          description: 有効期間の開始（含む）
          example: 2026-07-01T00:00:00Z
        validTo:
          type: string
          format: date-time
          description: 有効期間の終了（含む）
          example: 2026-07-31T23:59:59Z
        status:
          type: string
          default: ACTIVE
          description: 状態。SUSPENDED は緊急停止中（/coupons/suspended に載る）
          enum:
          - ACTIVE
          - SUSPENDED
      required:
      - couponId
      - distributionType
      - validFrom
      - validTo
    EligibilityRequest:
      type: object
      description: 該当判定（eligibility）リクエスト
      properties:
        attributes:
          type: object
          additionalProperties: {}
          description: 1会員分の属性。BFF がプロファイルストアから取得して渡す（本サービスは保持しない）。
          example:
            age: 25
            gender: female
            store: "1024"
        at:
          type: string
          format: date-time
          description: 判定基準時刻。省略時はサーバの現在時刻。
          example: 2026-07-07T09:00:00Z
      required:
      - attributes
    EligibilityResponse:
      type: object
      description: 該当判定（eligibility）レスポンス
      properties:
        coupons:
          type: array
          items:
            $ref: "#/components/schemas/EligibleCouponResponse"
      required:
      - coupons
    EligibleCouponResponse:
      type: object
      description: 該当クーポン（IDと配布種別のみ。詳細・画像・並び順の整形は BFF/CMS 側）
      properties:
        couponId:
          type: string
          description: クーポンID
          example: CP-2026-0001
        distributionType:
          type: string
          description: 配布種別（ALL または SEGMENT。INDIVIDUAL は eligibility の対象外）
          enum:
          - ALL
          - SEGMENT
          - INDIVIDUAL
      required:
      - couponId
      - distributionType
...